Extract rooms controller
