+++ app.server.js
@@ -1,9 +1,9 @@
 import express from "express";
-import fs from "fs";
 import path from "path";
 import favicon from "serve-favicon";
 import GoogleCredentialController from "./controllers/google.credentials.controller";
 import OfficeController from "./controllers/office.controller";
+import fetchRooms from "./controllers/rooms.controller";
 import Office from "./office.server";
 
 const ROOMS_SOURCE = process.env.ROOMS_SOURCE;
@@ -41,30 +41,6 @@ app.use(
 
 // FIX ME: here we have to get the google APIkey in another way.
 app.locals.googleCredential = new GoogleCredentialController(GOOGLECREDENTIAL);
-
-const fetchRooms = strategy => {
-  switch (strategy) {
-    //TODO add suport to fetch from endpoint
-    case "ENVIRONMENT":
-      return fetchFromEnvironment(process.env);
-    default:
-      return fetchFromFile();
-  }
-};
-
-const fetchFromFile = () => {
-  const roomsData = fs.readFileSync("./file/default.room.web.json");
-  const roomsDetail = JSON.parse(roomsData);
-
-  return new Promise((resolve, reject) => resolve(roomsDetail));
-};
-
-const fetchFromEnvironment = env => {
-  const roomsData = env.ROOMS_DATA;
-  const roomsDetail = JSON.parse(roomsData);
-
-  return new Promise((resolve, reject) => resolve(roomsDetail));
-};
 
 // routes
 app.get("/", (req, res) => {
